requirements.txt
Flask
scikit-learn
joblib
itsdangerous
PyJWT  
